uploading mothur log files from failed attempts at Lab9
I think the problem in this attempt is that with the screen.seqs() command it removed all but 1 sequence from the fasta. But also this was my first attempt and I was trying to figure out what to put for the fasta argument in the clearcut() function to create the .tre file. I think while trying to figure this out, I accidentally quit mother and then when I tried mothur/mothur again and then use current as the fasta argument, it didn't work because mothur didn't remember my files from before it quit. So this attempt failed. This is only part 1 of the log files. Parts 2 and 4 are just 'M', I'm not sure why, and parts 3 and 5 just show what was explained above.