uploading mothur log files from failed attempts at Lab9
This was my second attempt. The first mistake I think I made was when I ran the count.seqs() command after the unique.seqs() command. I used the same exact command from the Mac commands file, which gave the parameter name the file stability.trim.contigs.good.names in the count.seqs() command. This did not work because my code did not output a file named that. After a while and a lot of research I realized in the latest release I think they changed the names file. I think it uses the count file now, although I am still very unsure about that. But, when I ran the next command, get.current(), my output was the same as in the video, so I continued on. 
Then, the same thing happened that happened in my first attempt. screen.seqs() got rid of all but 1 sequence.
And I again, still hadn't figured out what to give the fasta parameter in clearcut() and probably accidentally quit mothur and thus ended this attempt