Add files via upload
This is the mothur log file for my fourth attempt.
I found the tutorial on mother's website and followed their steps in this file, which differed slightly from the instructions on Brightspace. First, it used mothur to create the stability.files. 
In the unique.seqs() command I copied and pasted, it passed the wrong name to fasta argument (and the same thing happened in remove.seqs() later on, highlighted in red circle, but then I fixed the argument passed to fasta)
The screen.seqs() command this time only removed 10 sequences, instead of all but 1 like in previous attempts
And after passing the wrong fasta file to fasta in many functions, I finally got it right, and got a .tre file. But I can't open it in fig tree. Get the error: Duplicate taxon M00967_