uploading mothur log files from failed attempts at Lab9
This is the mothur log file from my fifth attempt
This time I went back to following the instructions in brightspace
Made the same mistakes as before. screen.seqs() removed all but 1 sequence again.
Looking at the log file it looks like everything in the end worked? But it didn't create the .tre file. At least I can't find it in my files.